Reset all drag state when initializing a vertical tab drag
Previously, VerticalTabDragHandler expected the state to be in a clean
state on initialization, but this is not the case.

The horizontal tab strip also does not expect the state to be clean, so
this is matching this behavior. A TODO was created to understand why,
but for now this unblocks crashes.

Change-Id: I6c59605f5c2523fb6e3370f8bf85ad25dc520ae4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7274920
Commit-Queue: Kaan Alsan <alsan@chromium.org>
Reviewed-by: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1560734}