Emit ReadAnythingLifecycleObserver events for Immersive Reading Mode
These were mostly handled correctly, except for TabWillDetach. It was
relying on the WebUI to call CloseUI on the host
(ReadAnythingImmersiveWebView)- however, this doesn't work because we
remove the WebUI from the ReadAnythingImmersiveWebView before the WebUI
closes.

Therefore, this change calls ReadAnythingController->OnEntryHidden when
the WebUI is removed from the ReadAnythingImmersiveWebView instead of
relying on the WebUI's built in call to CloseUI().

Also adds tests to confirm all ReadAnythingLifecycleObserver are emitted
correctly.

Bug: 465446246
Change-Id: I0e5efa670e186034ee828e8ebfa9f9c7ec876966
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7272245
Commit-Queue: Eitan Goldberger <eitang@google.com>
Reviewed-by: Jacobo Misael Tapia <jatapiaro@google.com>
Reviewed-by: Shari Trewin <trewin@google.com>
Cr-Commit-Position: refs/heads/main@{#1560739}