Remove "kReduceUserAgentPlatformOsCpu", which is default-enabled.
This flag maps to the so-called "Phase 5" described in
https://www.chromium.org/updates/ua-reduction/#proposed-rollout-plan.

Given that it's been enabled for dozens of milestones at this point, we
are confident that it is web-compatible and no longer needed.

Low-Coverage-Reason: TRIVIAL_CHANGE Deleting a default-enabled feature, and code that tests both states.

Bug: 469455153
Change-Id: Idd3df6a5e42bfe925d36c82cfd6464a5f6eb3c06
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7266462
Reviewed-by: Victor Tan <victortan@chromium.org>
Owners-Override: Mike Taylor <miketaylr@chromium.org>
Commit-Queue: Mike Taylor <miketaylr@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1560817}