Port some ExtensionTabUtil tab group code to desktop Android, step 1
Specifically, GetWindowIdOfGroup() and CreateTabGroupObject(). These
are dependencies of the sessions API and tab groups API.

In order to get the window for a tab group, add ContainsTabGroup() to
the other tab group methods in TabListInterface. The new function is
not implemented yet on Android -- I'll add the JNI in a follow-up.

This is all just plumbing, no functional changes yet.

Bug: 430344931, 405219627, 405219902
Change-Id: I3528d8f9c7a8693ae11545c606e6c0134be150ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7271571
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Reviewed-by: Michael Wojcicka <mwoj@google.com>
Reviewed-by: Mike Wasserman <msw@chromium.org>
Commit-Queue: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1560820}