[CADisplayLink] Use a force_update flag when doing UpdateVSyncDisplay
The current implementation of
ExternalBeginFrameSourceMac::UpdateVSyncDisplay() updates the
DisplayLink by sequentially calling
ExternalBeginFrameSourceMac::SetVSyncDisplayID() with kInvalidDisplayId
and then the correct display ID to force an update. This method has a
drawback in that ExternalBeginFrameSourceMac::SetVSyncDisplayID()
generates an error message when called with kInvalidDisplayId, leading
to unnecessary error output and recording of failures in histograms. To
address this, a force_update flag is being added to SetVSyncDisplayID().

Bug: 345275139
Change-Id: I5403ed022fb13f01789bf6dd688decf5aae7f509
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7276489
Auto-Submit: Maggie Chen <magchen@chromium.org>
Commit-Queue: Avi Drissman <avi@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1560826}