Fix intent picker handling in WebAppIntegrationTestDriver
When calculating the state for a Browser window, we should use
the intent picker icon visibility for that window, and not the
state for whatever browser() happens to return at the time.

This revealed another issue where the browser() getter wasn't a
pure getter and in some situations would actually modify the state
of the browser by inserting a new blank tab. Before the AfterStateChange
code would always call browser(), so any changes this might make
would generally be taken care of by that. But with this fix browser()
is no longer called from there. Instead the logic to make sure a
browser has at least one tab is moved to the sync test code,
since that test harness is the only place where a browser is actually
created without any tabs.

This seems to fix some test flakiness, although there is still some
unrelated flakiness due to timeouts in CompletePageLoadForAllWebContents.

Bug: 458201369, 456952424
Change-Id: Ie864118e2e6606c6453c2cfc04c80dfa841ec224
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7275616
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1560860}