Add a feature flag to use the Rust Pix QR code validator
Currently, the payments driver performs two tests when checking for a Pix
QR code on the clipboard: the first is a fast test that is imprecise but
considered safe to implement in C++, while the second check actually
tries to parse it in C++ in the utility process.

This adds a feature flag to optionally enable the Rust validator: with
the Rust validator, all validation is in-process, so no quick but
imprecise test is needed. To support a gradual transition, this CL adds
a feature flag for using the Rust validator. When the feature flag is
enabled, the payments driver only uses the Rust validator. When the
feature flag is disabled, the payments driver uses the original C++
validator, but also performs an additional Rust validation so that the
results can be compared.

Finally, add additional metrics:
- The Rust validator has a more fine-grained set of results, so add a
  new histogram and enum definition when the Rust validator feature is
  enabled.
- The C++ validator also logs the result of Rust validation (using an
  approximate mapping of the Rust result to the original metrics enum)
  to validate that there are no significant differences between the two.

Bug: 454668204
Change-Id: I3c19baea20c591115148d168cb67ebeceeb486c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7081914
Reviewed-by: Olivia Saul <osaul@google.com>
Reviewed-by: Stephen McGruer <smcgruer@chromium.org>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Longsheng Yin <longsheng@google.com>
Cr-Commit-Position: refs/heads/main@{#1560903}