Changing Traffic Annotation CHECK back to DCHECK
As part of the implementation of the Proxy Override Feature, an existing
DCHECK was changed to CHECK in order to follow the latest Chromium
guidance. However, that change is causing crashes like these in M144
Beta:
https://crash.corp.google.com/browse?q=reportid=%278b2fcd7631227f23%27#4

Looking at the stack trace of the crash, it appears to be going through
a code path that predates my code changes. Meaning this DCHECK may have
been DCHECKing all along in non-Debug builds in the wild.

The mitigation, to unblock M144, is to revert it back to DCHECK. The
right fix will be to hunt down the cases where an invalid traffic
annotation is being set, but that will take more time.

Bug: 469006851
Change-Id: I9bd73b509b35eb9501d89aba1629f3431965d959
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7276860
Reviewed-by: mmenke <mmenke@chromium.org>
Commit-Queue: Sébastien Lalancette <seblalancette@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1560936}