[<install> Element] Do not show the permission prompt
Adds `InstallFromElement()` to the mojo interface that accepts the same
parameters as `Install()` but is used for installations triggered from
an <install> element. We do not want to show the permission prompt for
triggers from an <install> element so the permission is allowed by
default unless previously denied.

- Feature flag: InstallElement
- Internal design doc: https://docs.google.com/document/d/1rGvLhD4SR8Y9M1wVmqgyesPNkbZGU7HOqlttjEFJ5Vo/edit?tab=t.0
- Test site: https://kbhlee2121.github.io/pwa/web-install/index.html

- Explainer: https://github.com/WICG/install-element

Bug: 454827186, 462477497
Change-Id: I9fdda4b6c6eeb191a313dc935a03cd6dbe188beb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7266918
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: Kristin Lee <kristinlee@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1560950}