Release capture when start showing the multitask menu.
Since the tablet_mode_multitask_menu_controller is a PreTargetHandler,
it can continue handling the gesture even if someone else (a browser
window in this case) capture the window. This can cause an issue
because
1) the controller can consume the event.
2) which means the browser window will not receive the rest of gesture
  sequence, and will keep the capture.
3) the menu will not receive tap event because the capture is still
  owned by the browser window.

This CL fixes the issue by
1) forcibly releasing the capture when menu drag starts.
2) before releasing, the transfer gesture sequence to the
   menu's window so that it can continue processing the gesture.

I also fixed the NoCrashWhenDraggingSplitViewDivider test as this was
testing a code path that does not happn in production, and fixing it
caused the failure. (see the comment in the code)

Bug: 463884426
Test: covered by unit test

Change-Id: I940bf1cdb7163b5eec009db7316bf8d27c5c60b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7262106
Reviewed-by: Georg Neis <neis@chromium.org>
Commit-Queue: Mitsuru Oshima <oshima@chromium.org>
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1560967}