[contextual-search] Fix ContextualSearchSessionHandle ownership for Lens
For Lens flows, there could be a race condition in when the session
handle is set on the WebContents helper and when the WebUI controller is
initialized. In this CL, the Contextual Tasks WebUI Controller still
tries to take ownership of the session handle at initialization but also
preserves the lazy initialization fallback for the cases where the
session handle is provided late.

This change also renames GetContextualSessionHandle() to
GetOrCreateContextualSessionHandle() in all WebUI Controllers to more
explicitly indicate that the function may create a session handle if one
doesn't exist. It also adds TryTakeSessionHandleFromWebContents() helper
method to avoid code duplication in the constructor and
GetOrCreateContextualSessionHandle().

Bug: b:467511872
Change-Id: I909cb302b1c5e628aa07898bb4c8497739f9e5c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7275884
Commit-Queue: Moe Ahmadi <mahmadi@chromium.org>
Reviewed-by: Sophie Chang <sophiechang@chromium.org>
Reviewed-by: Juan Mojica <juanmojica@google.com>
Cr-Commit-Position: refs/heads/main@{#1560977}