[rustmojo] Represent sequences in rust
This CL contains the minimal code we need to use
`base::SequencedTaskRunner` from rust code. Doing so requires us to also
represent the `scoped_refptr` type, to hold the runners safely.

The key parts of the CL are:

1. Callbacks: this defines a closure type which we can send to C++ for
execution via `base::BindOnce`, since we can't directly create a
`base::OnceClosure` in rust. This is in a separate crate because GN
doesn't let us mix C++ and rust code in the same target, and the shim
functions rely on this type.

2. ScopedRefPtr: a direct analogue of the `scoped_refptr` type from C++.
It makes heavy use of the fact that C++ types are represented in rust
using opaque, zero-sized types, which have relaxed restrictions,
especially around aliasing. We define several traits in order to use
these safely.

3. SequencedTaskRunner: a direct analogue of a
`base::SequencedTaskRunner` in rust. We only port over the methods we
need, using C++ shims when necessary to make sure the wire types match
up. To hide the unsafety from users, we provide a single type that wraps
a `ScopedRefPtr` and provides convenient functions for getting the
default runner and posting tasks to it.

Bug: 444509367
Change-Id: I49cbb2799a81f4bfd7a510041cbf9a6255c1cf9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7251653
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Commit-Queue: Devon Loehr <dloehr@google.com>
Cr-Commit-Position: refs/heads/main@{#1560986}