[ios] Change the Logo & Doodle margins to reduce NTP load jank
This change adjusts the NTP logo and Doodle margins so that the Logo and
Doodle consume the same amount of space. This makes it so that when the
Doodle loads, the animation to display it does not cause the rest of the
NTP to be pushed down.

This is all behind a new flag: ConsistentLogoDoodleHeight.

Screencast without these adjustments:
http://screencast/cast/NjEyMzc1NDczMTA3NzYzMnxmYmExMWJjNS1kZg

Screencast with margin adjustments:
http://screencast/cast/NjM4NTM4MjA3NjcxMDkxMnxkNjI3NTBiMi1lYQ

Bug: 470056769
Change-Id: Iaca09a49d3cdbc05c0f0c258c723f7de11c2d74a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7276284
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Scott Yoder <scottyoder@google.com>
Cr-Commit-Position: refs/heads/main@{#1561105}