Optimize scroll-target-group using OrderedScopeTree
Improve the performance of scroll-target-group by replacing full DOM
traversals with an OrderedScopeTree-based management system.

Previously, any change to an anchor or a scroll-target-group property
triggered a global recalculation.

This patch integrates scroll-target-group into the OrderedScope
infrastructure, allowing for localized updates and efficient item
tracking.

Also, this patch modifies OrderedScope to allow to control if items of
the scope should be saved in the scope or in some user-controlled
storage (e.g. in this case items will be saved in focus_group_ of
ScrollMarkerGroupData), so that the OrderedScopeTree infrastructure can
be used just for scope management and batch updates.

Change-Id: I4927214e984f9f270f9ce46818157a0b8a9e7ca7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7264892
Commit-Queue: Daniil Sakhapov <sakhapov@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1561123}