[views-ax] Expose Webview in platform accessibility tree
This CL exposes the Webview in the platform accessibility tree when
ViewsAX is enabled. It gets rid of the hack we had in the
Webview::GetNativeViewAccessible override, where we'd replace the
Webview's accessible object by the one of the native view on Windows and
macOS.

Without this change, the root of the web content's AXTree isn't included
in the Views AXTree.

Previously, without ViewsAX, we had the following platform AX trees for the following Views hierarchy.

Views tree:
++MyContainerView
++++Webview (references a RenderWidgetHostView, which ultimately references a web content AX tree)

Windows (UIA) AX Tree:
++MyContainerView
++++AXFragmentRootPlatformNodeWin
++++++RootWebArea (start of the web content AX tree)

macOS AX Tree:
++MyContainerView
++++RenderWidgetHostViewCocoa
++++++RootWebArea (start of the web content AX tree)

AXFragmentRootPlatformNodeWin and RenderWidgetHostViewCocoa are two native accessible objects that must be inserted in the final accessibility tree exposed to Windows (UIA) and macOS.

Notice how the Webview view wasn't exposed in the final accessibility tree; this was caused by the override of Webview::GetNativeViewAccessible.

When enabling ViewsAX, the following errors would happen:
1. Browser crash in BrowserAccessibilityManager because the serialized node corresponding to Webview had both a child node and a child tree id, which is not allowed. We fix it with the change in view_accessibility.cc.
2. The native accessible object isn't included in the tree because the internal native object AXFragmentRootPlatformNodeWin and RenderWidgetHostViewCocoa's tree walk logic pointed to the parent of the Webview instead of the webview itself. This is fixed by the changes in webview.cc

Tested with Narrator, NVDA, JAWS, and VoiceOver. While there remains
several issues with the accessibility tree to fix with ViewsAX, this
change doesn't happen to break the major ATs anymore than it is. It also
successfully enables web content accessibility.

Bug: 40672441
Change-Id: I5e98537dd3e2d9cb3b2926309b44bec53bb8ec5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7261564
Reviewed-by: Jacques Newman <janewman@microsoft.com>
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Reviewed-by: Lucas Radaelli <lucasradaelli@google.com>
Auto-Submit: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Commit-Queue: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1561127}