Use SimpleURLLoader::DownloadHeadersOnly in components/metrics
This change replaces calls to SimpleURLLoader's
DownloadToStringOfUnboundedSizeUntilCrashAndDie with DownloadHeadersOnly
in various parts of the Chrome codebase. The body of the response was
not being used in these instances, so having SimpleURLLoader not store
the response body is more efficient and better conveys how the callers
interact with the request responses.

No major changes in functionality are expected as a result of this other
than SimpleURLLoader no longer saving the response bodies while the
request is being downloaded.

This CL was uploaded by git cl split.

Bug: None
Change-Id: I549e9b9bf8fdc03ac244be0fce1e61b2b0874490
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7270689
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Commit-Queue: Andrew Williams <awillia@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1561190}