[iOS] Fix TabGridTransition when selected cell is not visible
In some case isSelectedCellVisible and _tabGridCellItem nullity are out
of sync. As the only purpose of isSelectedCellVisible is to know if the
item can be used or not, remove the method and relies on the item value.

Fixed: 466347599
Change-Id: I5041c98b49c72b4e20df82725e4b5b4c2930280b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7280325
Commit-Queue: Ewann Pellé <ewannpv@chromium.org>
Reviewed-by: Ewann Pellé <ewannpv@chromium.org>
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1561199}