[Contextual Tasks] Push task info to webui when updated
This patch addresses a bug when opening a task in a tab from the side
panel. Previously we'd never update the task param on the WebUI URL
after initially being set. This caused an issue where the "open in tab"
functionality would stop working - The param in the URL is used in a
number of places, so it needs to be kept up to date.

This patch pushes the task ID, thread ID, and turn ID to the WebUI
whenever they are updated. Those values replace the existing ones in
the URL.

Bug: b:460833892, b:469717965
Change-Id: I9d24daebbc0a07d489be2b32bec16735c501ac80
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7271167
Commit-Queue: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Reviewed-by: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1561224}