[Mac][Omnibox WebUI] Create configurable tooltips acceptance
Currently on Mac, tooltips are suppressed if the window is not the Key
window. This prevents tooltips from appearing in the AIM popup which is
a TYPE_POPUP window (which never gets keyboard input).

This CL introduces an AcceptTooltipEvents configuration similar to the
AcceptMouseEvents pattern. This allows the bridge to query the delegate
for the right tooltip policy. ChromeRenderWidgetHostViewMacDelegate sets
the policy to allow tooltips for TopChrome WebUI URLS even if they are
not the key window.

Bug: 469864001
Change-Id: I4ea6e2cabc311e2acc4fc138522e6b6850583fc1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7277199
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Keren Zhu <kerenzhu@chromium.org>
Commit-Queue: David Yeung <dayeung@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1561227}