DOM: Fix aria-expanded state for menu items
The CL https://crrev.com/c/6642578 implemented aria-expanded integration
with menu items, as they are now command invokers and can invoke
popovers. The impl mostly lived in `AXNodeObject::IsExpanded()`, however
this week I caught a related bug. While expanded=true gets set when a
menu item triggers a popover via command, it does not get unset when the
popover is dismissed.

HTMLElement has a `MarkPopoverInvokersDirty()` method that gets
triggered in the popover dismissal path to mark all command invokers of
the freshly-dismissed popover as "dirty" with regards to accessibility,
so that their expanded state among other things gets updated.

Before this CL, this logic was tight enough as to only consider button
elements as command invokers, and this CL loosens this logic, and the
HTML element collection of invokers that a popover maintains, to include
menu item elements. This allows menu items to be marked as dirty and
have their expanded state updated in the buggy case.

R=masonf

Bug: 406566432, 425682466
Change-Id: I4304d1fbb078f29b6e11b37f4af35c9b5fed986a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7265521
Commit-Queue: Dominic Farolino <dom@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1561239}