[SLAMS] Enable SLAMS based on the single ClientSideDetectionServerModelForScamDetectionAndroid flag.
The ClientSideDetectionIntelligentScanDelegateAndroid now considers the
kClientSideDetectionServerModelForScamDetectionAndroid feature flag when
determining if intelligent scan is enabled and if scam warnings should
be shown. This allows SLAMS to be active even if the separate
kClientSideDetectionSendIntelligentScanInfoAndroid or
kClientSideDetectionShowScamVerdictWarningAndroid flags are not enabled.

This change will simplify the experiment setup while both CLAMS and
SLAMS are being rolled out. Otherwise, both experiments need to enable
the same flags, which will cause conflict.

Bug: 462643935
Change-Id: Ie9d5dc70eede963760e9f07571c234da005790a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7271104
Reviewed-by: Seung Jae Lim <andysjlim@chromium.org>
Commit-Queue: Xinghui Lu <xinghuilu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1561245}