[ntp-next] Do not retrieve non-contextual suggestions when sync off
Switches the client to "chrome-contextual" when there is delayed context
added and history sync is not enabled to ensure the suggestions
are only of contextual suggestion type. Also fixes an issue where
the number of files/context in the composebox is tracked wrong, causing
delayed context uploads with sync enabled to not request contextual
suggestions after non-delayed context is added and deleted.

Bug: 460596844
Change-Id: Ia2ae479c3262694b0ac8952d7b0e71120623236d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7276493
Reviewed-by: Riley Tatum <rtatum@google.com>
Commit-Queue: Jennifer Serrano <jennserrano@google.com>
Cr-Commit-Position: refs/heads/main@{#1561263}