infra: Add cronet's codesearch builder to the cronet rotation console
Since the codesearch builders are in the "infra" project, we can't
easily add them to the cronet gardening rotation.

Adding them to the rotation console is the next easiest thing.

crbug.com/406463490 is on file to move the codesearch builders to the
chromium project. Should be able to effectively revert this CL if/when
that's done.

Bug: 469715210
Change-Id: I6b76215f91ed355e38dfdcf0fa047748fa1f3fc0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7272084
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Reviewed-by: Stefano Duo <stefanoduo@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1561265}